Change back to write build methods on "Module" (rather than on separate "BeanFactory")
Removes SimpleFactoryWriter, write the module with all the build methods together on the same class again. We can do this because we get the JavaFileObject early
 - This means our Module classes are now visible/usable in src/main
 - We write the content on last round as we only have the full ordering at that point
